Time: 0 ms (100%), Space: 23 MB (97.42%) - LeetHub

+++ 3242-count-elements-with-maximum-frequency/3242-count-elements-with-maximum-frequency.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int maxFrequencyElements(vector<int>& nums) {
-        int freq[101] = {0};
-        int max = 0, res = 0;
+        uint8_t freq[101] = {0};
+        uint8_t max = 0, res = 0;
         for (int n : nums) {
-            int f = ++freq[n];
+            uint8_t f = ++freq[n];
             if (f > max) {
                 max = f;
                 res = f;
